requirements.txt
nicegui
fastapi
uvicorn
requests
groq
cerebras-cloud-sdk
matplotlib
seaborn
pandas
numpy
plotly
pytest
httpx
python-multipart